requirements.txt
anyio==3.6.2
audioread==3.0.0
boto3==1.26.142
botocore==1.29.142
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
fastapi==0.95.2
ffmpeg-python==0.2.0
future==0.18.3
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
jmespath==1.0.1
joblib==1.2.0
librosa==0.8.1
llvmlite==0.39.1
mido==1.2.10
numba==0.56.4
numpy==1.21.6
packaging==23.1
platformdirs==3.5.1
pooch==1.7.0
pretty-midi==0.2.10
pycparser==2.21
pydantic==1.10.8
pydub==0.25.1
pyFluidSynth==1.3.2
python-dateutil==2.8.2
requests==2.31.0
resampy==0.4.2
s3transfer==0.6.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
starlette==0.27.0
threadpoolctl==3.1.0
typing_extensions==4.6.2
urllib3==1.26.16
uvicorn==0.22.0
zipp==3.15.0
